experimental: create core plugin so it will be easy to create other data fetching layers on top of this idea
