+++ packages/example-app/package.json
@@ -15,7 +15,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "next": "^12.2.3",
+    "next": "^12.2.4",
     "next-swr-endpoints": "workspace:*",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
