+++ packages/next-api-endpoints-core-plugin/client/package.json
@@ -1,0 +1,7 @@
+{
+  "main": "../dist/client.js",
+  "browser": "../dist/client.mjs",
+  "module": "../dist/client.mjs",
+  "types": "../dist/client.d.ts",
+  "sideEffects": false
+}
