+++ packages/next-api-endpoints-core-plugin/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "next-swr-endpoints",
+  "name": "next-api-endpoints-core-plugin",
   "version": "1.0.0",
   "description": "",
   "main": "dist/index.js",
@@ -16,7 +16,7 @@
   "scripts": {
     "build": "tsup",
     "dev": "tsup --watch",
-    "test": "# vitest run"
+    "test": "vitest run"
   },
   "keywords": [],
   "author": "",
@@ -25,14 +25,20 @@
     "@types/prettier": "^2.6.4",
     "@types/react": "^18.0.10",
     "@types/react-dom": "^18.0.5",
+    "@types/webpack": "^5.28.0",
+    "@types/webpack-sources": "^3.2.0",
     "next": "^12.2.4",
-    "next-api-endpoints-core-plugin": "workspace:^1.0.0",
     "prettier": "^2.7.1",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
-    "swr": "^2.0.0-beta.6",
     "tsup": "^6.0.1",
     "typescript": "^4.7.2",
     "vitest": "^0.20.2"
+  },
+  "dependencies": {
+    "acorn": "^8.7.1",
+    "acorn-walk": "^8.2.0",
+    "astring": "^1.8.3",
+    "webpack-sources": "^3.2.3"
   }
 }
