+++ packages/next-api-endpoints-core-plugin/server/package.json
@@ -1,0 +1,6 @@
+{
+  "main": "../dist/server.js",
+  "module": "../dist/server.mjs",
+  "types": "../dist/server.d.ts",
+  "sideEffects": false
+}
