+++ packages/next-api-endpoints-core-plugin/tsup.config.ts
@@ -2,17 +2,15 @@ import { defineConfig } from "tsup";
 
 export default defineConfig({
   clean: true,
-  dts: {
-    resolve: [/next-api-endpoints-core-plugin/],
-  },
+  dts: true,
   entry: [
     "src/index.ts",
-    "src/swr-client-endpoint-loader.ts",
-    "src/swr-server-endpoint-loader.ts",
+    "src/client-endpoint-loader.ts",
+    "src/server-endpoint-loader.ts",
     "src/client.ts",
     "src/server.ts",
   ],
-  external: ["next/server", "swr", "swr/mutation"],
+  external: ["next/server", "next"],
   format: ["cjs", "esm"],
   minify: process.env.NODE_ENV === "production",
   sourcemap: true,
