+++ packages/next-swr-endpoints/src/index.ts
@@ -1,9 +1,8 @@
 import type { SWRResponse } from "swr";
 import type { SWRMutationResponse } from "swr/mutation";
-import type { NextConfig } from "next";
 import type { Parser } from "./parser";
-import type { Configuration } from "webpack";
-import type { HandlerCallback } from "./server";
+import type { HandlerCallback } from "next-api-endpoints-core-plugin/server";
+import { createPlugin } from "next-api-endpoints-core-plugin";
 
 export function query<Input, Output>(
   parser: Parser<Input>,
@@ -19,51 +18,10 @@ export function mutation<Input, Output>(
   throw new Error("This code path should not be reached");
 }
 
-export function withSwrApiEndpoints(given: NextConfig = {}): NextConfig {
-  const pageExtensions = (
-    given.pageExtensions ?? ["js", "jsx", "ts", "tsx"]
-  ).flatMap((value) => {
-    return [value, `api.${value}`, `swr.${value}`];
-  });
-
-  const escapedPageExtensions = pageExtensions.map((x) =>
-    x.replace(/\\./g, "\\.")
-  );
-  const testRegex = new RegExp(
-    `\\.(api|swr)\\.(${escapedPageExtensions.join("|")})$`
-  );
-
-  return {
-    ...given,
-    webpack(config: Configuration, context) {
-      config.module?.rules?.unshift({
-        test: testRegex,
-        issuerLayer: { or: ["api", "middleware"] },
-        use: [
-          {
-            loader: "next-swr-endpoints/server-loader",
-            options: { nextRuntime: context.nextRuntime },
-          },
-          context.defaultLoaders.babel,
-        ],
-      });
-      config.module?.rules?.unshift({
-        test: testRegex,
-        issuerLayer: { not: { or: ["api", "middleware"] } },
-        use: [
-          {
-            loader: "next-swr-endpoints/client-loader",
-            options: {
-              projectDir: context.dir,
-              pageExtensionsRegex: testRegex,
-              basePath: context.config.basePath,
-            },
-          },
-          context.defaultLoaders.babel,
-        ],
-      });
-      return given.webpack ? given.webpack(config, context) : config;
-    },
-    pageExtensions,
-  };
-}
+export const withSwrApiEndpoints = /*#__PURE__*/ createPlugin({
+  capturedExtensions: ["swr"],
+  clientLoaderPath: "next-swr-endpoints/client-loader",
+  serverLoaderPath: "next-swr-endpoints/server-loader",
+  clientPackageName: "next-swr-endpoints/client",
+  serverPackageName: "next-swr-endpoints/server",
+});
